data packing lab

+++ labs/memory_bound/data_packing/init.cpp
@@ -8,7 +8,7 @@ S create_entry(int first_value, int second_value) {
   entry.i = first_value;
   entry.s = static_cast<short>(second_value);
   entry.l = static_cast<long long>(first_value * second_value);
-  entry.d = static_cast<double>(first_value) / maxRandom;
+  entry.d = static_cast<float>(first_value) / maxRandom;
   entry.b = first_value < second_value;
 
   return entry;
